Fixed the build using the new type of permissions and refactored the testing helpers
The test.helpers is now a seperate project (because of circular dependancies otherwise)
The permissions.perm now generates the package permissions. This is not complete yet and the service permissions should be generated as well. BND does not support this automatically right now, so this has to be done by hand.
Also, all version numbers are 2.0.0 for consistency. Or should we simply call this the 1.0.0 release?