Refactored the websockets implementation
It now uses the servlet and proxy concepts correctly
Also removed using the configuration admin to create agent proxies, they can be created directly and just registered when the remote client has connected. After the client has been connected, it will register the proxy as a service and the session manager will then connect it to the rest of the cluster. This seems to work perfectly in the tests and the code now is a lot more understandable.
Also includes a hack to make sure no ClassNotFoundException is thrown by the Jetty codebase