Improved synchronization in the runtime environment
Added a SessionManagerTests that adds and removes lots of agents in lots of threads in high speed. This caused lots of Exceptions in the previous commit.
Added its own lock object in the BaseAgent that improves locking behavior (especially when someone tries to extend it with their own synchronized methods).
Added locking in the SessionImpl to make sure no updates can be sent until the session is really fully setup.
Added checks for disconnecting in the SessionImpl, it will now ignore updates when a disconnect has been called. Updates could still come from a matcher, because the map with all the session won't update fast enough over several thread (ask me if you want more explenation of this behavior).
Minimized the logging for the tests (otherwise the Jenkins build server would probably flip)