First draft of the proposal to add a Status object to the Agent
This can solve many of the concurrency issues we are having. This assumes a getStatus() method that will return a snapshot of the current connection state of the Agent. This way another thread can not change any of the parameters, because you have a local copy. Also in the SessionImpl the bids and prices that are being sent are ignored when a disconnect has occured in the meantime.

Right now all the tests seem to succeed, but further tests are needed.